refactor: Signup step2 구성요소 리팩토링
