+++ src/features/signup/ui/steps/step-2/CollegeSelect.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useSignupStore } from '../../model/signup.store';
+import { useSignupStore } from '@/features/signup/model';
 
-export default function CollegeSelect() {
+export function CollegeSelect() {
   const { college, setProfile } = useSignupStore();
 
   return (
@@ -12,7 +12,7 @@ export default function CollegeSelect() {
         value={college}
         onChange={(e) => setProfile({ college: e.target.value })}
         placeholder="ex) 공과대"
-        className="h-14 w-full rounded-full border border-black px-4 text-base outline-none focus:bg-gray-50"
+        className="h-14 w-full rounded-full border-2 border-black px-4 text-base outline-none focus:bg-gray-50"
       />
     </div>
   );
