+++ src/features/signup/ui/steps/step-2/index.ts
@@ -1,0 +1,4 @@
+export { NicknameForm } from './NicknameForm';
+export { AgeStudentIdForm } from './AgeStudentIdForm';
+export { CollegeSelect } from './CollegeSelect';
+export { GenderForm } from './GenderForm';
